Changed "Renderer.enableSecondaryPalette" to use a u16 parameter to allow for larger screen heights

+++ Oxygen/oxygenengine/source/oxygen/simulation/bindings/RendererBindings.cpp
@@ -352,7 +352,7 @@ namespace
 		RenderParts::instance().getPaletteManager().writePaletteEntryPacked(0, index, color);
 	}
 
-	void Renderer_enableSecondaryPalette(uint8 line)
+	void Renderer_enableSecondaryPalette(uint16 line)
 	{
 		RenderParts::instance().getPaletteManager().setPaletteSplitPositionY(line);
 	}
